+++ composer/algorithms/__init__.py
@@ -55,7 +55,7 @@ from composer.algorithms.layer_freezing import LayerFreezing
 from composer.algorithms.mixup import MixUp
 from composer.algorithms.no_op_model import NoOpModel
 from composer.algorithms.progressive_resizing import ProgressiveResizing
-from composer.algorithms.randaugment import RandAugment, RandAugmentTransform
+from composer.algorithms.rand_augment import RandAugment, RandAugmentTransform
 from composer.algorithms.sam import SAM
 from composer.algorithms.selective_backprop import SelectiveBackprop
 from composer.algorithms.seq_length_warmup import SeqLengthWarmup
