+++ docs/source/method_cards/rand_augment.md
@@ -1,0 +1,1 @@
+../../../composer/algorithms/rand_augment/README.md
