--- docs/source/method_cards/randaugment.md
@@ -1,1 +1,0 @@
-../../../composer/algorithms/randaugment/README.md
