+++ docs/source/tables/algorithms_table.md
@@ -21,7 +21,7 @@ Name|Functional|Attribution|tl;dr|Math-Equivalent
 [LayerFreezing](https://github.com/mosaicml/composer/tree/dev/composer/algorithms/layer_freezing)|`cf.freeze_layers`|[Many; Raghu et al, 2017](https://arxiv.org/abs/1706.05806)|Progressively freezes layers during training|❌
 [MixUp](https://github.com/mosaicml/composer/tree/dev/composer/algorithms/mixup)|`cf.mixup_batch`|[Zhang et al, 2017](https://arxiv.org/abs/1710.09412)|Blends pairs of examples and labels|❌
 [ProgressiveResizing](https://github.com/mosaicml/composer/tree/dev/composer/algorithms/progressive_resizing)|`cf.resize_batch`|[Fast AI](https://github.com/fastai/fastbook/blob/780b76bef3127ce5b64f8230fce60e915a7e0735/07_sizing_and_tta.ipynb)|Scale the input image size over the course of training|❌
-[RandAugment](https://github.com/mosaicml/composer/tree/dev/composer/algorithms/randaugment)|`cf.randaugment_image`|[Cubuk et al, 2020](https://openaccess.thecvf.com/content_CVPRW_2020/html/w40/Cubuk_Randaugment_Practical_Automated_Data_Augmentation_With_a_Reduced_Search_Space_CVPRW_2020_paper.html)|Applies a series of random augmentations|❌
+[RandAugment](https://github.com/mosaicml/composer/tree/dev/composer/algorithms/rand_augment)|`cf.randaugment_image`|[Cubuk et al, 2020](https://openaccess.thecvf.com/content_CVPRW_2020/html/w40/Cubuk_Randaugment_Practical_Automated_Data_Augmentation_With_a_Reduced_Search_Space_CVPRW_2020_paper.html)|Applies a series of random augmentations|❌
 [SAM](https://github.com/mosaicml/composer/tree/dev/composer/algorithms/sam)||[Foret et al, 2021](https://arxiv.org/abs/2010.01412)|SAM optimizer measures sharpness of optimization space|❌
 [ScaleSchedule](https://github.com/mosaicml/composer/tree/dev/composer/algorithms/scale_schedule)||[Many](https://www.mosaicml.com)|Scale the learning rate schedule by a factor|❌
 [SelectiveBackprop](https://github.com/mosaicml/composer/tree/dev/composer/algorithms/selective_backprop)|`cf.selective_backprop`|[Jiang et al, 2019](https://arxiv.org/abs/1910.00762)|Drops examples with small loss contributions|❌
